able to handle default: cases of a switch properly:
each default case is a branch registered in the BranchPool just like the usual Branch objects for normal cases in a switch
the BranchInstrumentation has been extended to add instrumentation so the ExecutionTracer is called properly for the branch of the default case as explained in BranchInstrumentation.getSwitchInstrumentation.

next: Branch should no longer extend from BytecodeInstruction!