Rename methods to avoid confusion about terminology
Some of the method names can be misleading because there is an ambiguity
in the terminology: "object" vs. "Object". For example, the method name
"attemptObjectGeneration" suggests that the method is responsible for
creating a new instance of a given class (i.e., an object). However,
in reality, it is specifically intended to find instances that match the
type java.lang.Object.

Signed-off-by: Sebastian Schweikl <schweikl@posteo.de>