Eliminate unsafe casts and raw types using generics and adapter pattern
The root of many of those issues seems to be located in the MOSA
classes, spreading further to any other class that somehow comes into
touch with MOSA. MOSA evolves test chromosomes but is used in a context
where test suite chromosomes are expected. AbstractMOSA and its
subclasses are in fact already generified, but they use an upper type
bound that is too general (T extends Chromosome, but if anything,
TestChromosome would be the only reasonable upper bound) and circumvent
the type system using raw types and unchecked casts/assignments. For
example, in DynaMOSA, T is instantiated with the concrete
TestSuiteChromosome type (to make it work in environments that expect
such chromosomes), but after creating an instance of the genetic
algorithm, the chromosome factory is changed to one that produces
test chromosomes instead (because MOSA was never designed to evolve
entire test suites, just simple test cases). This is only possible
because Java supports raw types for the sake of backwards compatibility.
To solve these problems, we make use of the adapter design pattern.
There is now an adapter class that wraps MOSA-type algorithms (and thus
evolves TestChromosomes) but returns the results as
TestSuiteChromosomes. Essentially, whenever a test suite is required,
we take the best test cases found by MOSA and artificially add them to
a newly created TestSuiteChromosome, which is then returned. This idea
is in fact identical to what has been done before, but now it is done
in a type-safe manner.

Furthermore, the GeneticAlgorithm class has been added another type
parameter, allowing us to specify which kind of fitness functions the
algorithm works with. This is esp. useful in the context of MOSA, which
does not work with ordinary FitnessFunction<TestChromosome> but requires
a special subclass thereof, namely TestFitnessFunction. Before this
change, it was necessary to unsafely down-cast every
FitnessFunction<TestChromosome> to TestFitnessFunction in the code.
This is now no longer necessary because the type system already
ensures that no fitness functions other than TestFitnessFunction are
used with MOSA.

Signed-off-by: Sebastian Schweikl <schweikl@posteo.de>