DynaMOSA: fix fitness evaluation bug and implement failure penalties for infeasible goals
MultiCriteriaManager evaluates the population against the set of
currently "active" goals (i.e. those goals without direct control
dependencies to any other still uncovered goal). During fitness
evaluation, the previous implementation modified the set of current
goals on the fly, while fitness evaluation of the current generation
still took place. That is, once it was found that a previously
uncovered goal was now covered by the given test case, the goal was
immediately removed from the set of current goals. Therefore, this
goal was no longer taken into account during the fitness evaluation
of the remaining test cases of the current population. This is a
problem, since we could potentially miss a shorter test case that might
cover the same goal.
The fix removes the newly covered goals from the set of current goals
only after every single individual of the current population has been
evaluated.

This commit also includes the implementation of failure penalties.
Whenever a test case fails to cover its intended goal, the failure
penalty for that goal is increased. Once it becomes too high, we
deem the goal as being too hard to cover and give up on it. That is, we
remove this yet uncovered goal from the set of current goals and never
touch it again.

Signed-off-by: Sebastian Schweikl <schweikl@posteo.de>