Cloning a statement now also resets its TTL
Currently, a statement is cloned if we want to use it as part of
another test case, for example when we perform crossover of test cases.
In the new test case, the statement should be treated as "freshly
generated", i.e., it should be treated the same way as if it was
inserted by the mutation operator. As such, it's former TTL should no
longer matter in the context of the new test case.

Signed-off-by: Sebastian Schweikl <schweikl@posteo.de>