Don't reset TTL of statements when being cloned
This effectively reverts commit 11b421c39e4da68180acd862b98b0c49e5d4c4e8

Resetting the TTLs of statements when being cloned makes sense only in
some cases (e.g., when a statement is to be inserted into another test
case where it effectively acts as a new statement), whereas in other
cases it doesn't (e.g., when the statement is cloned during mutation to
make a defensive copy upon which the mutations will be performed). For
now, we never reset the TTL when a statement is cloned, and don't
handle the cases where it would make sense to reset the TTL differently.
For this, we would have to make the method resetTTL() public again, and
then manually call it after every clone() operation whenever we see fit.

Signed-off-by: Sebastian Schweikl <schweikl@posteo.de>