Change logger message from warn to debug level
It happens quite frequently that a method was not registered during
method dependence analysis. This is completely intentional as we're
only interested in analyzing the classes on the project class path.
Otherwise, we would have to analyze the entire JDK. The message
should therefore not be logged at the warn level, since this clutters
the logs considerably.

Signed-off-by: Sebastian Schweikl <schweikl@posteo.de>