Added new property METHOD_DEPENDENCE_ANALYSIS
If enabled, EvoSuite performs a dependence analysis of the methods
in the class path. More specifically, this allows us to identify
data dependencies between methods that stem from writing to and
subsequently reading from a common attribute of a class. As the
writing method performs a side effect that potentially influences the
reading method, it is desirable to test such pairs of methods together.

Signed-off-by: Sebastian Schweikl <schweikl@posteo.de>