Fix: ignoring bad offspring
In the old implementation, tests with internal exception or that reached the timeout were still included in th offspring population, although with a very bad objectives scores. Due to GA randomness, these ad offspring may survive and leaad to evaluating similar tests in the next iteration. Now, bad offspring tests are simply not added in the offspring population