Sbst2019 fixes (#287)
* Fixing errors related to very short search-budgets

* Removing unecessary log messages

* Fix: ignoring bad offspring

In the old implementation, tests with internal exception or that reached the timeout were still included in th offspring population, although with a very bad objectives scores. Due to GA randomness, these ad offspring may survive and leaad to evaluating similar tests in the next iteration. Now, bad offspring tests are simply not added in the offspring population

* Adding extra checks

Just added extra checks in DynaMOSA to guarantee that the search stop (ASAP) when the seach budget is over

Co-authored-by: apanichella <a.panichella@tudelft.nl>