Added Improvements to the exploration algorithm cache (#22)
# Details
The optimizations for path skipping and constraint caching were not accurate.

# Changes

## Already Seen Paths
Fixed path divergence check. Now after we checked for a divergence we compare against the previously explored paths for skipping in case of a divergence.

## Constraint Caching
We were wrongly using counterexample cache, now we use a few elements (see the [chapter two](https://www.overleaf.com/read/zvdnjhxmkckn) of the thesis report for more info).