Fix bug in results output locations
The previous directory structure was not unique enough to make sure that
each executions has its own file to write to. It was writing all classes
corresponding to the same project in the same file. This commit fixes
this by introducing one more layer in the directory structure for the
classes.

It also cleans up some of the naming