Add possibility for overriding the class path
Currently the script determines the classpth based on the jars present
in the individual project folder. With this change it allows the user to
override this class path by providing a file, called 'CLASSPATH' within
this directory. The first line of this file will be used as the class
path.